requirements.txt
beautifulsoup4==4.12.3
mysql-connector-python==9.0.0
phpserialize==1.3
PyYAML==6.0.1
requests==2.32.3
requests-cache==1.2.1
selenium==4.22.0
typer==0.12.3
python-dotenv==1.0.1